requirements.txt
Flask==3.0.2
Werkzeug==3.0.1
web3==6.15.1
python-dotenv==1.0.1
Pillow==10.2.0  # For image handling
requests==2.31.0  # For API calls 